+++ tutorials/30-communication-and-bbs/60-middleware-ping/ping.cpp
@@ -13,7 +13,7 @@ public:
    void run () {
 	long pingCnt = 0;
 	TIME_LOOP(0, 1*SECONDS) {
-	  PRINTF("ping %ld at %lld\n", pingCnt, NOW());
+	  PRINTF("ping %ld at %lld\n", pingCnt, static_cast<long long>(NOW()));
           tSend = NOW();
 	  ping.publish(pingCnt);
           pingCnt++;
@@ -23,7 +23,7 @@ public:
    long put([[gnu::unused]] const long topicId, [[gnu::unused]] const long len, void* data, [[gnu::unused]] const NetMsgInfo& netMsgInfo) {
         tReceive = NOW();
         long cnt = *(long*)data;
-         PRINTF("pong %ld at %lld, dif = %6.3f us\n", cnt, tReceive, (double)(tReceive - tSend)/MICROSECONDS);
+         PRINTF("pong %ld at %lld, dif = %6.3f us\n", cnt, static_cast<long long>(tReceive), (double)(tReceive - tSend)/MICROSECONDS);
          return 1;
    }
